rename transformerTemplate to specify data type
