+++ grasshopper-loader.js
@@ -66,10 +66,11 @@ function processData(err, fileName, stream, cb){
   try{
     transformer = getTransformer(fileName, cb)
   }catch(err){
+    console.log("transformer error", err, cb);
     if(cb) return cb(err);
     throw err;
   }
-
+  
   console.log("Streaming %s to elasticsearch.", fileName);
 
   return pipeline(fileName, stream, transformer, cb);
