+++ lib/pointTransformer.js
