+++ transformers/arkansas.js
@@ -1,3 +1,3 @@
 'use strict';
-var template = require('../lib/transformerTemplate');
+var template = require('../lib/pointTransformer');
 module.exports = template('ADR_LABEL', 'ADR_CITY', 'ADR_STATE', 'ADR_ZIP5')
