+++ transformers/maine.js
@@ -1,3 +1,3 @@
 'use strict';
-var template = require('../lib/transformerTemplate');
+var template = require('../lib/pointTransformer');
 module.exports = template('ADDRESS', 'MUNICIPALI', 'STATE', 'ZIPCODE');
