+++ transformers/new_york.js
@@ -1,3 +1,3 @@
 'use strict';
-var template = require('../lib/transformerTemplate');
+var template = require('../lib/pointTransformer');
 module.exports = template('AddressLabel', 'CityTownName', 'State', 'ZipCode');
