+++ transformers/north_carolina.js
@@ -1,3 +1,3 @@
 'use strict';
-var template = require('../lib/transformerTemplate');
+var template = require('../lib/pointTransformer');
 module.exports = template('COMP_STR', '', 'STATE_NAME', 'ZIP_CODE');
